pkcs2pem: Output a single file with the private key and certificate.
(Actually, it already did, but it also gave you a separate certificate
file.)

It’s more convenient to deal with a single file.  Many programs accept
it as the cert file without needing to specify a separate key file.
For other programs, you can just pass the same file as both the cert
file and the key file.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>