newline-fix.bash: Force the bash prompt to start after a newline
Some commands don’t print a newline at the end of their output, or
take long enough to return that you’ve already typed part of the
next command.  Either causes bash to start its prompt in the middle
of a line, which confuses it.  This script sets $PROMPT_COMMAND to
echo a magic sequence of terminal commands that will display a red
“<no LF>\n” before the prompt if it would otherwise start in the
middle of a line.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>