programming: Sync gccrun and disasm from my locker (oops)
disasm gains a -o option to provide arguments to objdump. The argument
to -o gets interpreted by the shell, allowing multiple arguments to be
passed to objdump.

gccrun now passes parameters past the first to gcc, requiring you to
quote the code to be compiled. It also allows the wrapper to be
interpreted by the shell, allowing a wrapper with arguments.